requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic[email]==2.9.2
pydantic-settings==2.6.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
